Enable hashed passwords with PowerShell
- PR #929 introduced the ability to generate passwords for the
gateway admin user, this is a good step forwards for Windows
users. It did introduce an inconsistency in the format that
passwords are stored by not using a form of hashing. Hashing
of secrets is used extensively within OpenFaaS/OpenFaaS Cloud
whether with Swarm or Kubernetes via helm. If there are
concerns about using a hashed value for a password I would
suggest raising an issue to track this and have any decision
we make applied for all users (not just PowerShell users).

As a  compromise I've introduced hashing by default and added
a new flag called -noHash which can be used to replicate the
behaviour of the original PR.

After feedback from other contributors I also looked into whether
the flag syntax could match the existing syntax but left this as
is. Bash will use --no-auth and PowerShell will use --noAuth.

This was tested on Docker Swarm on Windows.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>