Consolidate deploy_stack.sh files
Currently there are multiple deploy_stack.sh files and the user has to select the version appropriate to the target architecture so that the correct images are deployed.  Further, the armhf deploy script lacks the basic auth by default scripting that is in the x86_64 version.

This change adds a case/select to deploy_stack.sh that will determine the hardware architecture and select the appropriate docker-compose file.

The deploy_stack.armhf.sh is also changed, rather than deleted, so that existing references are still valid.

Signed-off-by: Richard Gee <richard@technologee.co.uk>